+++ phovea.js
@@ -10,19 +10,19 @@ module.exports = function(registry) {
 
   registry.push('autoload', 'caleydo_links', function() { return System.import('./src/link_renderer'); }, {});
 
-  registry.push('link-representation', 'link-block', function() { return System.import('./src/link_representation'); }, {
+  registry.push('link-representation', 'link-block', function() { return System.import('./src/LinkRepresentation'); }, {
   'name': 'Block Link Representation',
   'factory': 'createBlockRep',
   'granularity': 0
  });
 
-  registry.push('link-representation', 'link-group', function() { return System.import('./src/link_representation'); }, {
+  registry.push('link-representation', 'link-group', function() { return System.import('./src/LinkRepresentation'); }, {
   'name': 'Group Link Representation',
   'factory': 'createGroupRep',
   'granularity': 5
  });
 
-  registry.push('link-representation', 'link-item', function() { return System.import('./src/link_representation'); }, {
+  registry.push('link-representation', 'link-item', function() { return System.import('./src/LinkRepresentation'); }, {
   'name': 'Item Link Representation',
   'factory': 'createItemRep',
   'granularity': 10
