+++ src/index.ts
@@ -2,9 +2,8 @@ export * from './d3util';
 export * from './databrowser';
 export * from './layout_d3util';
 export * from './link_renderer';
-export * from './link_representation';
+export * from './LinkRepresentation';
 export * from './link';
-export {default as LinkContainer} from './link';
-export * from './parser';
+export * from './Parser';
 export * from './selectioninfo';
-export {default as bind} from './tooltip';
+export * from './ToolTip';
